+++ c-v/src/components/Footer/DeleteAllButton.jsx
@@ -1,17 +1,7 @@
 import { styled } from 'styled-components';
+
 import { useDeleteConfirmModalStore } from '../../stores/ModalStore';
 import { useListStore } from '../../stores/ListStore';
-
-const StyledDeleteAllButton = styled.button`
-  background: none;
-  color: #dd959a;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  &:hover {
-    color: #d5525b;
-  }
-`;
 
 const DeleteAllButton = () => {
   const openDeleteConfirmModal = useDeleteConfirmModalStore(
@@ -28,5 +18,16 @@ const DeleteAllButton = () => {
     </StyledDeleteAllButton>
   );
 };
+
+const StyledDeleteAllButton = styled.button`
+  background: none;
+  color: #dd959a;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  &:hover {
+    color: #d5525b;
+  }
+`;
 
 export default DeleteAllButton;
