+++ c-v/src/components/Footer/Footer.jsx
@@ -2,13 +2,6 @@ import { styled } from 'styled-components';
 
 import DeleteAllButton from './DeleteAllButton';
 
-const FooterContainer = styled.div`
-  width: 100%;
-  height: 5%;
-  display: flex;
-  justify-content: center;
-`;
-
 const Footer = () => {
   return (
     <FooterContainer>
@@ -16,5 +9,12 @@ const Footer = () => {
     </FooterContainer>
   );
 };
+
+const FooterContainer = styled.div`
+  width: 100%;
+  height: 5%;
+  display: flex;
+  justify-content: center;
+`;
 
 export default Footer;
