+++ c-v/src/components/Main/Command.jsx
@@ -3,60 +3,13 @@ import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import dragDropIcon from '../../assets/images/drag-drop-icon.svg';
-import deleteIcon from '../../assets/images/delete-white.svg';
 import { useListStore } from '../../stores/ListStore';
 import { useEditCommandModalStore } from '../../stores/ModalStore';
 import useCommandStore from '../../stores/CommandStore';
 import deleteCommand from '../../api/deleteCommand';
 
-const CommandWrapper = styled.li`
-  width: 250px;
-  ${({ $isDragging }) => $isDragging && 'opacity: 0.5'}
-  position: relative;
-  z-index: 99;
-`;
-
-const CommandHeader = styled.div`
-  position: relative;
-  z-index: 0;
-  width: 100%;
-  padding: 2px 3% 2px 3%;
-  display: flex;
-  background-color: #aeaeae;
-  justify-content: space-between;
-  border-radius: 5px 5px 0 0;
-  border-top: 1px solid #d9d9d9;
-  border-left: 1px solid #d9d9d9;
-  border-right: 1px solid #d9d9d9;
-
-  > h5 {
-    color: #fff;
-    font-size: 12px;
-  }
-  > img {
-    cursor: pointer;
-  }
-`;
-
-const CommandContent = styled.div`
-  width: 100%;
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid ${({ $isOver }) => ($isOver ? '#00B3FF' : '#d9d9d9')};
-  border-radius: 0 0 5px 5px;
-  > p {
-    width: 80%;
-    font-size: 11px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    text-align: center;
-    cursor: pointer;
-  }
-`;
+import dragDropIcon from '../../assets/images/drag-drop-icon.svg';
+import deleteIcon from '../../assets/images/delete-white.svg';
 
 const Command = ({ listItem }) => {
   const {
@@ -122,5 +75,53 @@ const Command = ({ listItem }) => {
     </CommandWrapper>
   );
 };
+
+const CommandWrapper = styled.li`
+  width: 250px;
+  ${({ $isDragging }) => $isDragging && 'opacity: 0.5'}
+  position: relative;
+  z-index: 99;
+`;
+
+const CommandHeader = styled.div`
+  position: relative;
+  z-index: 0;
+  width: 100%;
+  padding: 2px 3% 2px 3%;
+  display: flex;
+  background-color: #aeaeae;
+  justify-content: space-between;
+  border-radius: 5px 5px 0 0;
+  border-top: 1px solid #d9d9d9;
+  border-left: 1px solid #d9d9d9;
+  border-right: 1px solid #d9d9d9;
+
+  > h5 {
+    color: #fff;
+    font-size: 12px;
+  }
+  > img {
+    cursor: pointer;
+  }
+`;
+
+const CommandContent = styled.div`
+  width: 100%;
+  height: 50px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid ${({ $isOver }) => ($isOver ? '#00B3FF' : '#d9d9d9')};
+  border-radius: 0 0 5px 5px;
+  > p {
+    width: 80%;
+    font-size: 11px;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    text-align: center;
+    cursor: pointer;
+  }
+`;
 
 export default Command;
