+++ c-v/src/components/Main/NewCommandButton.jsx
@@ -1,17 +1,7 @@
 import styled from 'styled-components';
+
 import { useAddCommandModalStore } from '../../stores/ModalStore';
 import { useListStore } from '../../stores/ListStore';
-
-const StyledNewCommandButton = styled.button`
-  background: none;
-  color: #c6c6c6;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s ease;
-  &:hover {
-    color: #6a6a6a;
-  }
-`;
 
 const NewCommandButton = () => {
   const openAddCommandModal = useAddCommandModalStore(
@@ -28,5 +18,16 @@ const NewCommandButton = () => {
     </StyledNewCommandButton>
   );
 };
+
+const StyledNewCommandButton = styled.button`
+  background: none;
+  color: #c6c6c6;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  &:hover {
+    color: #6a6a6a;
+  }
+`;
 
 export default NewCommandButton;
