+++ c-v/src/components/Modal/EditList/EditInput.jsx
@@ -1,15 +1,6 @@
 import { useState } from 'react';
 import styled from 'styled-components';
 
-const StyledEditInput = styled.input`
-  width: 125px;
-  height: 18px;
-  border: 0.3px solid #414141;
-  border-radius: 3px;
-  padding: 2px 0 2px 5px;
-  font-size: 12px;
-`;
-
 const EditInput = ({ value, onChange }) => {
   const [inputValue, setInputValue] = useState(value);
 
@@ -25,5 +16,14 @@ const EditInput = ({ value, onChange }) => {
     ></StyledEditInput>
   );
 };
+
+const StyledEditInput = styled.input`
+  width: 125px;
+  height: 18px;
+  border: 0.3px solid #414141;
+  border-radius: 3px;
+  padding: 2px 0 2px 5px;
+  font-size: 12px;
+`;
 
 export default EditInput;
