+++ c-v/src/components/header/List.jsx
@@ -1,7 +1,31 @@
 import styled from 'styled-components';
+import { useEffect, useState } from 'react';
 
 import { useListStore } from '../../stores/ListStore';
-import { useEffect, useState } from 'react';
+import useGetList from '../../hooks/useGetList';
+
+const List = () => {
+  const setListName = useListStore((state) => state.setListName);
+  const [updatedList, setUpdatedList] = useState([]);
+  const { list, isSuccess } = useGetList();
+
+  useEffect(() => {
+    if (isSuccess) setUpdatedList(list);
+  }, [list, isSuccess]);
+
+  const clickHandler = (event) => {
+    const selectName = event.target.innerText;
+    setListName(selectName);
+  };
+
+  return (
+    <ListWrapper onClick={clickHandler}>
+      {updatedList.map((listItem, index) => (
+        <li key={index}>{listItem.name}</li>
+      ))}
+    </ListWrapper>
+  );
+};
 
 const ListWrapper = styled.ul`
   position: absolute;
@@ -21,29 +45,5 @@ const ListWrapper = styled.ul`
     cursor: pointer;
   }
 `;
-
-const List = () => {
-  const setListName = useListStore((state) => state.setListName);
-  const [list, setList] = useState([]);
-
-  useEffect(() => {
-    chrome.runtime
-      .sendMessage({ type: 'get-list' })
-      .then((response) => setList(response.listData));
-  }, []);
-
-  const clickHandler = (event) => {
-    const selectName = event.target.innerText;
-    setListName(selectName);
-  };
-
-  return (
-    <ListWrapper onClick={clickHandler}>
-      {list.map((listItem, index) => (
-        <li key={index}>{listItem.name}</li>
-      ))}
-    </ListWrapper>
-  );
-};
 
 export default List;
