+++ c-v/src/components/header/ModifyButtons.jsx
@@ -1,8 +1,22 @@
 import styled from 'styled-components';
+
 import {
   useAddListModalStore,
   useEditListModalStore,
 } from '../../stores/ModalStore';
+
+const ModifyButtons = () => {
+  const openAddModal = useAddListModalStore((state) => state.openModal);
+  const openEditListModal = useEditListModalStore((state) => state.openModal);
+
+  return (
+    <ModifyButtonContainer>
+      <button onClick={openAddModal}>Add</button>
+      <span> / </span>
+      <button onClick={openEditListModal}>Edit</button>
+    </ModifyButtonContainer>
+  );
+};
 
 const ModifyButtonContainer = styled.div`
   display: flex;
@@ -24,18 +38,5 @@ const ModifyButtonContainer = styled.div`
     }
   }
 `;
-
-const ModifyButtons = () => {
-  const openAddModal = useAddListModalStore((state) => state.openModal);
-  const openEditListModal = useEditListModalStore((state) => state.openModal);
-
-  return (
-    <ModifyButtonContainer>
-      <button onClick={openAddModal}>Add</button>
-      <span> / </span>
-      <button onClick={openEditListModal}>Edit</button>
-    </ModifyButtonContainer>
-  );
-};
 
 export default ModifyButtons;
